app1/requirements.txt
Django==1.9.4
gunicorn==19.6.0
djangorestframework
markdown
django-filter